feat: action mapping
