+++ src/metastable/paths/__init__.py
@@ -1,0 +1,14 @@
+from metastable.paths.data_structures import IndexPair, BistableBoundaries, Sweeps
+from metastable.paths.bistable_region import get_bistable_epsilon_range, get_bistable_kappa_range
+from metastable.paths.sweep_generation import generate_sweep_index_pairs
+from metastable.paths.map_paths import map_switching_paths
+
+__all__ = [
+    'IndexPair',
+    'BistableBoundaries',
+    'Sweeps',
+    'get_bistable_epsilon_range',
+    'get_bistable_kappa_range',
+    'generate_sweep_index_pairs',
+    'map_switching_paths',
+]
